+++ RELEASE_NOTES.md
@@ -22,6 +22,13 @@ _20 April 2016_
 ## Version 1.1.0
 _12 April 2016_
 
+### New Features
+* Added EHR Certification ID widget to fold out on Search page
+* Added 'Cert ID' button to product search result rows
+
+## Next version
+_Date TBD_
+
 ### New features
 * Allow editing of new style CHPL Product Number
 * Show attestation status for each developer/acb in 'developer codes' table